+++ src/pydrobert/param/serialization.py
@@ -45,9 +45,7 @@ from ._classic_serialization import (
     DefaultTupleSerializer,
     deserialize_from_dict,
     deserialize_from_ini,
-    deserialize_from_json_to_dict,
     deserialize_from_json,
-    deserialize_from_yaml_to_dict,
     deserialize_from_yaml,
     JSON_STRING_DESERIALIZER_DICT,
     JSON_STRING_SERIALIZER_DICT,
@@ -71,8 +69,6 @@ from ._classic_serialization import (
     ParamConfigDeserializer,
     ParamConfigSerializer,
     ParamConfigTypeError,
-    serialize_from_dict_to_json,
-    serialize_from_dict_to_yaml,
     serialize_to_dict,
     serialize_to_ini,
     serialize_to_json,
@@ -82,6 +78,12 @@ from ._serializer import (
     RecklessJsonSerialization,
     register_reckless_json,
     unregister_reckless_json,
+)
+from ._file_serialization import (
+    deserialize_from_json_to_obj,
+    deserialize_from_yaml_to_obj,
+    serialize_from_obj_to_json,
+    serialize_from_obj_to_yaml,
 )
 from .config import YAML_MODULE_PRIORITIES  # deprecated
 
@@ -116,9 +118,9 @@ __all__ = [
     "DefaultTupleSerializer",
     "deserialize_from_dict",
     "deserialize_from_ini",
-    "deserialize_from_json_to_dict",
+    "deserialize_from_json_to_obj",
     "deserialize_from_json",
-    "deserialize_from_yaml_to_dict",
+    "deserialize_from_yaml_to_obj",
     "deserialize_from_yaml",
     "JSON_STRING_DESERIALIZER_DICT",
     "JSON_STRING_SERIALIZER_DICT",
@@ -144,8 +146,8 @@ __all__ = [
     "ParamConfigTypeError",
     "RecklessJsonSerialization",
     "register_reckless_json",
-    "serialize_from_dict_to_json",
-    "serialize_from_dict_to_yaml",
+    "serialize_from_obj_to_json",
+    "serialize_from_obj_to_yaml",
     "serialize_to_dict",
     "serialize_to_ini",
     "serialize_to_json",
